==> 2-data cleaning trust level with all hospitals.ipynb <==
import pandas as pd

# Using double backslashes
df = pd.read_excel("/Users/e410675/Library/CloudStorage/OneDrive-LondonSouthBankUniversity/LSBU - Jose's Collaborative Projects/Norfolk and Waveney CHPPD/2025 November/Data Engineering/Data/1-NHS England trust level CHPPD - data combined.xlsx")

df

# Group by 'Organisation Code' and find the most frequent 'Organisation Name' for each code
standard_names = (df.groupby(['Organisation Code', 'Organisation Name'])
                  .size()
                  .reset_index(name='count')
                  .sort_values(by=['Organisation Code', 'count'], ascending=[True, False])
                  .drop_duplicates(subset=['Organisation Code'])
                  .set_index('Organisation Code')['Organisation Name'])

# Create a mapping dictionary from the standard names
name_mapping = standard_names.to_dict()

# Replace the 'Organisation Name' column with the standardized names
df['Organisation Name'] = df['Organisation Code'].map(name_mapping)

# Create a new column for 'Year' and 'Month' extracted from the 'year_month' column
df['Year'] = pd.to_datetime(df['year_month']).dt.year
df['Month'] = pd.to_datetime(df['year_month']).dt.month

# If you want a single column for Year and Month as "YYYY-MM"
df['Year_Month'] = pd.to_datetime(df['year_month']).dt.to_period('M')

# Save the cleaned dataset or display it
df.to_excel(r"/Users/e410675/Library/CloudStorage/OneDrive-LondonSouthBankUniversity/LSBU - Jose's Collaborative Projects/Norfolk and Waveney CHPPD/2025 November/Data Engineering/Data/2-clean_data_trust_level with all hospitals.xlsx", index=False)  # Optional: save to a new Excel file
